UDP comm not working
cmds go out to SymNet but are not acknowkedged; however async updates from symnet are returned, but oddly enough, with a NAK\r appended.

Not sure about SU/SUQ command; probably should be 1, for use with CSG command.

This system used to work, so i'm not sure what's wrong with the hardware or config...network related?