fixed hh:mm overflow time display bug (?; testme); added VTP test file
VTP test file is for TSW-1050 that crashes on latest VTP project version.  Problem seems to be in the new History page, probably the overload of HTML strings via serial joins.  needs more testing and then workarounds to main VTP file.
BTW, VTP file works fine as an XPanel, but not on the actual TSW-1050 unit.